requirements.txt
Django==1.10.6
channels
asgi_redis
uber_rides
djangorestframework
wikipedia
requests
argparse
nltk
numpy
textblob
geopy
